turning everything into reuseable functions
so the code ran fine, but i would have to specify each
key.value for every line, this is not suitable when it comes to the
next categories, or when some values do not exist, or have a value of
n/a or *

The object is made like this

object = {
	key.value,
	key2.value
}

eg this would be
{
	illumination: value,
	lens: value,
	...
	all the way down to
	power: value,
	op_temp: value
}

This would be fine, if, we were not separating into categories. I want
(and finance wants) this to be automated, to minimise general grunt time
for myself, so I am spending the time to separate and automate
filtering of unwanted 'n/a' or '*' values, for each category.